refactor(protocol-designer): move AdvancedSettingsConfirmationModal to components
